Refactor catalog utils - copy miranda structure stuff - parse with parse
