+++ xscen/catalog.py
@@ -1,37 +1,27 @@
-# noqa: D100
+"""Catalog objects and related tools."""
 import ast
 import itertools
 import json
 import logging
 import os
 import re
-import subprocess
 import warnings
 from collections.abc import Mapping, Sequence
 from copy import deepcopy
-from datetime import datetime
-from glob import glob
-from pathlib import Path, PosixPath
-from typing import Any, Dict, List, Optional, Tuple, Union
+from pathlib import Path
+from typing import Any, Optional, Union
 
-import cftime
-import dask
 import fsspec as fs
 import intake_esm
-import netCDF4
 import numpy as np
 import pandas as pd
 import tlz
 import xarray
 import xarray as xr
-import yaml
-from dask.diagnostics import ProgressBar
-from intake.source.utils import reverse_format
 from intake_esm.cat import ESMCatalogModel
 
-from .config import CONFIG, args_as_str, parse_config, recursive_update
-from .io import get_engine
-from .utils import CV, ensure_correct_time, standardize_periods  # noqa
+from .config import CONFIG, args_as_str, recursive_update
+from .utils import CV, date_parser, ensure_correct_time, standardize_periods  # noqa
 
 logger = logging.getLogger(__name__)
 # Monkey patch for attribute names in the output of to_dataset_dict
@@ -44,10 +34,7 @@ __all__ = [
     "ID_COLUMNS",
     "ProjectCatalog",
     "concat_data_catalogs",
-    "date_parser",
     "generate_id",
-    "parse_directory",
-    "parse_from_ds",
     "unstack_id",
 ]
 
@@ -814,672 +801,6 @@ def concat_data_catalogs(*dcs):
     newcat._dependent_variables = dependent_variables
     newcat._requested_variable_freqs = requested_variable_freqs
     return newcat
-
-
-@parse_config
-def _get_asset_list(root_paths, globpat="*.nc", parallel_depth=1, compute=True):
-    """List files fitting a given glob pattern from a list of paths.
-
-    Search is done with GNU's `find` and parallelized through `dask`.
-
-    Parameters
-    ----------
-    root_paths: Sequence of strings or paths
-    """
-    if isinstance(root_paths, str):
-        root_paths = [root_paths]
-
-    # Support for wildcards in root_paths
-    root_star = deepcopy(root_paths)
-    for r in root_star:
-        if "*" in r:
-            new_roots = [
-                x
-                for x in glob(str(r))
-                if Path(x).is_dir() and Path(x).suffix != ".zarr"
-            ]
-            root_paths.remove(r)
-            root_paths.extend(new_roots)
-
-    @dask.delayed
-    def _file_dir_files(directory, pattern):
-        try:
-            if "/" in pattern:
-                *foldparts, extension = pattern.split("/")
-                foldpatt = "/".join(foldparts)
-                folders = subprocess.run(
-                    [
-                        "find",
-                        "-L",
-                        directory.as_posix(),
-                        "-type",
-                        "d",
-                        "-wholename",
-                        foldpatt,
-                    ],
-                    capture_output=True,
-                    encoding="utf-8",
-                ).stdout.split()
-                output = []
-                for folder in folders:
-                    output.extend(
-                        subprocess.run(
-                            ["find", "-L", folder, "-perm", "-o+r", "-name", extension],
-                            capture_output=True,
-                            encoding="utf-8",
-                        ).stdout.split()
-                    )
-            else:
-                output = subprocess.run(
-                    [
-                        "find",
-                        "-L",
-                        directory.as_posix(),
-                        "-perm",
-                        "-o+r",
-                        "-name",
-                        pattern,
-                    ],
-                    capture_output=True,
-                    encoding="utf-8",
-                ).stdout.split()
-        except Exception:
-            output = []
-        return output
-
-    files = list()
-    for r in root_paths:
-        root = Path(r)
-        dirs = []
-        for d in range(1, parallel_depth + 1):
-            pattern = "*/" * d
-            for x in root.glob(pattern):
-                if x.is_dir() and x.suffix != ".zarr" and x.parent.suffix != ".zarr":
-                    if x.parent in dirs:
-                        dirs.remove(x.parent)
-                    dirs.append(x)
-
-        files.extend(
-            [(root, _file_dir_files(directory, globpat)) for directory in dirs]
-        )
-
-        # add files in the first directory, unless the glob pattern includes
-        # folder matching, in which case we don't want to risk recomputing the same parsing.
-        if "/" not in globpat:
-            files.append((root, [str(x.absolute()) for x in root.glob(f"{globpat}")]))
-
-    if compute:
-        with ProgressBar():
-            (files,) = dask.compute(files)
-    return files
-
-
-def _name_parser(
-    path,
-    root,
-    patterns,
-    read_from_file=None,
-    attrs_map: Mapping[str, Any] = None,
-    xr_open_kwargs: Mapping[str, Any] = None,
-    logr: logging.Logger = None,
-):
-    """Extract metadata information from the file path.
-
-    Parameters
-    ----------
-    path : str
-        Full file path.
-    patterns : list or str
-        List of patterns to try in `reverse_format`, special wildcards to manage underscores are added.
-        See :py:func:`parse_directory`.
-    columns : list of string, optional
-        If given, the metadata is restricted to fields from this list.
-    read_from_file : list of string or dict, optional
-        If not None, passed directly to :py:func:`parse_from_ds` as `names`.
-    attrs_map : dict, optional
-        If `read_from_file` is not None, passed directly to :py:func:`parse_from_ds`.
-    xr_open_kwargs : dict, optional
-        If required, arguments to send to xr.open_dataset() when opening the file to read the attributes.
-
-    Returns
-    -------
-    dict
-        The metadata fields parsed from the path using the first fitting pattern.
-        If no pattern matched, {'path': None} is returned.
-    """
-    abs_path = Path(path)
-    path = abs_path.relative_to(root)
-    xr_open_kwargs = xr_open_kwargs or {}
-
-    d = {}
-    for pattern in patterns:
-        if len(Path(pattern).parts) != len(path.parts):
-            continue
-        try:
-            patt = pattern.replace("*", "_?")
-            d = reverse_format(patt, str(path))
-            fields_with_illegal = {
-                k: v
-                for k, v in d.items()
-                if ("_" in v and not k.startswith("_")) or "/" in v
-            }
-            if d and not fields_with_illegal:
-                break
-            else:
-                d = {}
-        except ValueError:
-            continue
-        except IndexError as err:
-            logger.debug(f"Filename parsing failed with {err} for {pattern} on {path}.")
-            continue
-    if not d:
-        logger.debug(f"No pattern matched with path {path}")
-        return {"path": None}
-
-    logger.debug(f"Parsed file path {path} and got {len(d)} fields.")
-
-    # files with a single year/month
-    if "date_end" not in d and "date_start" in d:
-        d["date_end"] = d["date_start"]
-
-    d["path"] = abs_path
-    d["format"] = path.suffix[1:]
-
-    if read_from_file:
-        try:
-            fromfile = parse_from_ds(
-                abs_path, names=read_from_file, attrs_map=attrs_map, **xr_open_kwargs
-            )
-        except Exception as err:
-            logger.error(f"Unable to parse file {path}, got : {err}")
-        else:
-            d.update(fromfile)
-
-    # strip to clean off lost spaces and line jumps
-    # do not include wildcarded fields
-    return {
-        k if not k.startswith("_") else k[1:]: v.strip() if isinstance(v, str) else v
-        for k, v in d.items()
-        if "?" not in k
-    }
-
-
-@parse_config
-def parse_directory(
-    directories: list,
-    globpattern: str,
-    patterns: list,
-    *,
-    id_columns: list = None,
-    read_from_file: Union[
-        bool,
-        Sequence[str],
-        tuple[Sequence[str], Sequence[str]],
-        Sequence[tuple[Sequence[str], Sequence[str]]],
-    ] = False,
-    homogenous_info: dict = None,
-    cvs: Union[str, PosixPath, dict] = None,
-    xr_open_kwargs: Mapping[str, Any] = None,
-    parallel_depth: int = 1,
-    only_official_columns: bool = True,
-) -> pd.DataFrame:
-    r"""Parse files in a directory and return them as a pd.DataFrame.
-
-    Parameters
-    ----------
-    directories : list
-        List of directories to parse. The parse is recursive and accepts wildcards (*).
-    globpattern : str
-        A glob pattern for file name matching, usually only a suffix like "*.nc".
-        May include folder matching, in which case don't forget that the search is parallelized for
-        subfolders up to the depth given by `parallel_depth`. Contrary to real posix glob patterns,
-        this makes no difference between "**" and "*".
-    patterns : list
-        List of possible patterns to be used by intake.source.utils.reverse_filename() to decode the file names. See Notes below.
-    id_columns : list
-        List of column names on which to base the dataset definition. Empty columns will be skipped.
-        If None (default), it uses :py:data:`ID_COLUMNS`.
-    read_from_file : boolean or set of strings or tuple of 2 sets of strings.
-        If True, if some fields were not parsed from their path, files are opened and
-        missing fields are parsed from their metadata, if found.
-        If a sequence of column names, only those fields are parsed from the file, if missing.
-        If False (default), files are never opened.
-        If a tuple of 2 lists of strings, only the first file of groups defined by the
-        first list of columns is read and the second list of columns is parsed from the
-        file and applied to the whole group.
-        It can also be a list of those tuples.
-    homogenous_info : dict, optional
-        Using the {column_name: description} format, information to apply to all files.
-    cvs: str or PosixPath or dict, optional
-        Dictionary with mapping from parsed name to controlled names for each column.
-        May have an additional "attributes" entry which maps from attribute names in the files to
-        official column names. The attribute translation is done before the rest.
-        In the "variable" entry, if a name is mapped to None (null), that variable will not be listed in the catalog.
-    xr_open_kwargs: dict
-        If needed, arguments to send xr.open_dataset() when opening the file to read the attributes.
-    parallel_depth: int
-        The level at which to parallelize the file search. A value of 1 (default and minimum), means the subfolders
-        of each directory are searched in parallel, a value of 2 would search the subfolders' subfolders in parallel, and so on.
-    only_official_columns: bool
-        If True (default), this ensures the final catalog only has the columns defined in :py:data:`COLUMNS`. Other fields in the patterns will raise an error.
-        If False, the columns are those used in the patterns and the homogenous info. In that case, the column order is not determined.
-        Path, format and id are always present in the output.
-
-    Notes
-    -----
-    - Offical columns names are controlled and ordered by :py:data:`COLUMNS`:
-        ["id", "type", "processing_level", "mip_era", "activity", "driving_institution", "driving_model", "institution",
-         "source", "bias_adjust_institution", "bias_adjust_project","experiment", "member",
-         "xrfreq", "frequency", "variable", "domain", "date_start", "date_end", "version"]
-    - Not all column names have to be present, but "xrfreq" (obtainable through "frequency"), "variable",
-        "date_start" and "processing_level" are necessary for a workable catalog.
-    - 'patterns' should highlight the columns with braces.
-        This acts like the reverse operation of `format()`. Fields will match alphanumeric parts of the path,
-        excluding the "_" and "/" characters. Prefix tokens can be used to change the behaviour:
-        "_" will also match underscores.
-        "?" will match normally, but will not be included in the output.
-        "\*" is the equivalent of "_?", it will match underscores and be excluded from the output.
-        Due to the limited implementation of this mini syntax, "\*" elements should only appear at the end of a file or folder name .
-        Example: `"{_source}/{?ignored project name}_{?}_{domain}_{variable}_{date_start}_{activity}_{experiment}_{processing_level}_{\*gibberish}.nc"`
-        Here, "source" will be the full folder name, even if it includes underscores. The two first sections of the filename will be excluded from the output,
-        the first one was given a name (ignore project name) to make the pattern readable. The last sections of the filenames ("\*gibberish") will be excluded.
-
-    Returns
-    -------
-    pd.DataFrame
-        Parsed directory files
-    """
-    homogenous_info = homogenous_info or {}
-    xr_open_kwargs = xr_open_kwargs or {}
-    pattern_fields = {
-        f if not f.startswith("_") else f[1:]
-        for f in set.union(*map(lambda p: set(re.findall(r"{([\w]*)}", p)), patterns))
-    }
-    if only_official_columns:
-        columns = set(COLUMNS) - homogenous_info.keys()
-        unrecognized = pattern_fields - set(COLUMNS)
-        if unrecognized:
-            raise ValueError(
-                f"Patterns include fields which are not recognized by xscen : {unrecognized}. "
-                "If this is wanted, pass only_official_columns=False to remove the check."
-            )
-
-    read_file_groups = False  # Whether to read file per group or not.
-    if not isinstance(read_from_file, bool) and not isinstance(read_from_file[0], str):
-        # A tuple of 2 lists
-        read_file_groups = True
-        if isinstance(read_from_file[0][0], str):
-            # only one grouping
-            read_from_file = [read_from_file]
-    elif read_from_file is True:
-        # True but not a list of strings
-        read_from_file = columns
-
-    if cvs is not None:
-        if not isinstance(cvs, dict):
-            with open(cvs) as f:
-                cvs = yaml.safe_load(f)
-        attrs_map = cvs.pop("attributes", {})
-    else:
-        attrs_map = {}
-
-    # Find files (returns a list of tuples of (root, delayed_file_list)).
-    files = _get_asset_list(
-        directories, globpat=globpattern, parallel_depth=parallel_depth, compute=False
-    )
-
-    # Paths is a delayed object
-    @dask.delayed
-    def _wrap_name_parser(paths, root, *args, **kwargs):
-        return [_name_parser(path, root, *args, **kwargs) for path in paths]
-
-    parsed = [
-        _wrap_name_parser(
-            paths,
-            root,
-            patterns,
-            read_from_file=read_from_file if not read_file_groups else None,
-            attrs_map=attrs_map,
-            xr_open_kwargs=xr_open_kwargs,
-        )
-        for root, paths in files
-    ]
-
-    with ProgressBar():  # Finding the files and parsing the names is done here.
-        (parsed,) = dask.compute(parsed)
-    parsed = list(itertools.chain(*parsed))
-
-    if not parsed:
-        raise ValueError("No files found.")
-    else:
-        logger.info(f"Found and parsed {len(parsed)} files.")
-
-    # Path has become NaN when some paths didn't fit any passed pattern
-    df = pd.DataFrame(parsed).dropna(axis=0, subset=["path"])
-
-    if only_official_columns:  # Add the missing official columns
-        for col in set(COLUMNS) - set(df.columns):
-            df[col] = None
-
-    # Parse attributes from one file per group
-    def read_first_file(grp, cols):
-        fromfile = parse_from_ds(grp.path.iloc[0], cols, attrs_map, **xr_open_kwargs)
-        logger.info(f"Got {len(fromfile)} fields, applying to {len(grp)} entries.")
-        out = grp.copy()
-        for col, val in fromfile.items():
-            for i in grp.index:  # If val is an iterable we can't use loc.
-                out.at[i, col] = val
-        return out
-
-    if read_file_groups:
-        for group_cols, parse_cols in read_from_file:
-            df = (
-                df.groupby(group_cols)
-                .apply(read_first_file, cols=parse_cols)
-                .reset_index(drop=True)
-            )
-
-    # Add homogeous info
-    for key, val in homogenous_info.items():
-        df[key] = val
-
-    # Replace entries by definitions found in CV
-    if cvs:
-        # Read all CVs and replace values in catalog accordingly
-        df = df.replace(cvs)
-        if "variable" in cvs:
-            # Variable can be a tuple, we still want to replace individual names through the cvs
-            df["variable"] = df.variable.apply(
-                lambda vs: vs
-                if isinstance(vs, str) or pd.isnull(vs)
-                else tuple(
-                    cvs["variable"].get(v, v)
-                    for v in vs
-                    if cvs["variable"].get(v, v) is not None
-                )
-            )
-
-    # translate xrfreq into frequencies and vice-versa
-    if {"xrfreq", "frequency"}.issubset(df.columns):
-        df["xrfreq"].fillna(
-            df["frequency"].apply(CV.frequency_to_xrfreq, default=pd.NA), inplace=True
-        )
-        df["frequency"].fillna(
-            df["xrfreq"].apply(CV.xrfreq_to_frequency, default=pd.NA), inplace=True
-        )
-
-    # Parse dates
-    if "date_start" in df.columns:
-        df["date_start"] = df["date_start"].apply(date_parser)
-    if "date_end" in df.columns:
-        df["date_end"] = df["date_end"].apply(date_parser, end_of_period=True)
-
-    # Checks
-    if {"date_start", "date_end", "xrfreq", "frequency"}.issubset(df.columns):
-        # All NaN dates correspond to a fx frequency.
-        invalid = (
-            df.date_start.isnull()
-            & df.date_end.isnull()
-            & (df.xrfreq != "fx")
-            & (df.frequency != "fx")
-        )
-        n = invalid.sum()
-        if n > 0:
-            warnings.warn(
-                f"{n} invalid entries where the start and end dates are Null but the frequency is not 'fx'."
-            )
-            logger.debug(f"Paths: {df.path[invalid].values}")
-            df = df[~invalid]
-
-    # todo
-    # - Vocabulary check on xrfreq and other columns
-    # - Format is understood
-
-    # Create id from user specifications
-    df["id"] = generate_id(df, id_columns)
-
-    # ensure path is a string
-    df["path"] = df.path.apply(str)
-
-    # Sort columns and return
-    if only_official_columns:
-        return df.loc[:, COLUMNS]
-    return df
-
-
-def parse_from_ds(
-    obj: Union[os.PathLike, xr.Dataset],
-    names: Sequence[str],
-    attrs_map: Optional[Mapping[str, str]] = None,
-    **xrkwargs,
-):
-    """Parse a list of catalog fields from the file/dataset itself.
-
-    If passed a path, this opens the file.
-
-    Infers the variable from the variables.
-    Infers xrfreq, frequency, date_start and date_end from the time coordinate if present.
-    Infers other attributes from the coordinates or the global attributes. Attributes names
-    can be translated using the `attrs_map` mapping (from file attribute name to name in `names`).
-
-    If the obj is the path to a Zarr dataset and none of "frequency", "xrfreq", "date_start" or "date_end"
-    are requested, :py:func:`parse_from_zarr` is used instead of opening the file.
-    """
-    get_time = bool(
-        {"frequency", "xrfreq", "date_start", "date_end"}.intersection(names)
-    )
-    if not isinstance(obj, xr.Dataset):
-        obj = Path(obj)
-
-    if isinstance(obj, Path) and obj.suffixes[-1] == ".zarr" and not get_time:
-        logger.info(f"Parsing attributes from Zarr {obj}.")
-        ds_attrs, variables = _parse_from_zarr(obj, get_vars="variable" in names)
-        time = None
-    elif isinstance(obj, Path) and obj.suffixes[-1] == ".nc":
-        logger.info(f"Parsing attributes with netCDF4 from {obj}.")
-        ds_attrs, variables, time = _parse_from_nc(
-            obj, get_vars="variable" in names, get_time=get_time
-        )
-    else:
-        if isinstance(obj, Path):
-            logger.info(f"Parsing attributes with xarray from {obj}.")
-            obj = xr.open_dataset(obj, engine=get_engine(obj), **xrkwargs)
-        ds_attrs = obj.attrs
-        time = obj.indexes["time"] if "time" in obj else None
-        variables = set(obj.data_vars.keys()).difference(
-            [v for v in obj.data_vars if len(obj[v].dims) == 0]
-        )
-
-    rev_attrs_map = {v: k for k, v in (attrs_map or {}).items()}
-    attrs = {}
-
-    for name in names:
-        if name == "variable":
-            attrs["variable"] = tuple(sorted(variables))
-        elif name in ("frequency", "xrfreq") and time is not None and time.size > 3:
-            # round to the minute to catch floating point imprecision
-            freq = xr.infer_freq(time.round("T"))
-            if freq:
-                if "xrfreq" in names:
-                    attrs["xrfreq"] = freq
-                if "frequency" in names:
-                    attrs["frequency"] = CV.xrfreq_to_frequency(freq)
-            else:
-                warnings.warn(
-                    f"Couldn't infer frequency of dataset {obj if not isinstance(obj, xr.Dataset) else ''}"
-                )
-        elif name in ("frequency", "xrfreq") and time is None:
-            attrs[name] = "fx"
-        elif name == "date_start" and time is not None:
-            attrs["date_start"] = time[0]
-        elif name == "date_end" and time is not None:
-            attrs["date_end"] = time[-1]
-        elif name in rev_attrs_map and rev_attrs_map[name] in ds_attrs:
-            attrs[name] = ds_attrs[rev_attrs_map[name]].strip()
-        elif name in ds_attrs:
-            attrs[name] = ds_attrs[name].strip()
-
-    logger.debug(f"Got fields {attrs.keys()} from file.")
-    return attrs
-
-
-def _parse_from_zarr(path: os.PathLike, get_vars=True):
-    """Obtain the list of variables and the list of global attributes from a zarr dataset, reading the JSON files directly.
-
-    Variables are those
-    - where .zattrs/_ARRAY_DIMENSIONS is not empty
-    - where .zattrs/_ARRAY_DIMENSIONS does not contain the variable name
-    - who do not appear in any "coordinates" attribute.
-    """
-    path = Path(path)
-
-    if (path / ".zattrs").is_file():
-        with (path / ".zattrs").open() as f:
-            ds_attrs = json.load(f)
-    else:
-        ds_attrs = {}
-
-    variables = []
-    if get_vars:
-        coords = []
-        for varpath in path.iterdir():
-            if varpath.is_dir() and (varpath / ".zattrs").is_file():
-                with (varpath / ".zattrs").open() as f:
-                    var_attrs = json.load(f)
-                if (
-                    varpath.name in var_attrs["_ARRAY_DIMENSIONS"]
-                    or len(var_attrs["_ARRAY_DIMENSIONS"]) == 0
-                ):
-                    coords.append(varpath.name)
-                if "coordinates" in var_attrs:
-                    coords.extend(
-                        list(map(str.strip, var_attrs["coordinates"].split(" ")))
-                    )
-        variables = [
-            varpath.name
-            for varpath in path.iterdir()
-            if varpath.name not in coords and varpath.is_dir()
-        ]
-    return ds_attrs, variables
-
-
-def _parse_from_nc(path: os.PathLike, get_vars=True, get_time=True):
-    """Obtain the list of variables, the time coordinate, and the list of global attributes from a netCDF dataset, using netCDF4."""
-    ds = netCDF4.Dataset(str(path))
-    ds_attrs = {k: ds.getncattr(k) for k in ds.ncattrs()}
-
-    variables = []
-    if get_vars:
-        coords = []
-        for name, var in ds.variables.items():
-            if "coordinates" in var.ncattrs():
-                coords.extend(
-                    list(map(str.strip, var.getncattr("coordinates").split(" ")))
-                )
-            if len(var.dimensions) == 0 or name in var.dimensions:
-                coords.append(name)
-        variables = [var for var in ds.variables.keys() if var not in coords]
-
-    time = None
-    if get_time and "time" in ds.variables:
-        time = xr.CFTimeIndex(
-            cftime.num2date(
-                ds["time"][:], calendar=ds["time"].calendar, units=ds["time"].units
-            ).data
-        )
-    ds.close()
-    return ds_attrs, variables, time
-
-
-def date_parser(
-    date,
-    *,
-    end_of_period: bool = False,
-    out_dtype: str = "period",
-    strtime_format: str = "%Y-%m-%d",
-    freq: str = "H",
-) -> Union[str, pd.Period, pd.Timestamp]:
-    """Return a datetime from a string.
-
-    Parameters
-    ----------
-    date : str
-        Date to be converted
-    end_of_period : bool, optional
-        If True, the date will be the end of month or year depending on what's most appropriate
-    out_dtype : str, optional
-        Choices are 'period', 'datetime' or 'str'
-    strtime_format : str, optional
-        If out_dtype=='str', this sets the strftime format
-    freq : str
-        If out_dtype=='period', this sets the frequency of the period.
-
-    Returns
-    -------
-    pd.Period, pd.Timestamp, str
-        Parsed date
-    """
-    # Formats, ordered depending on string length
-    fmts = {
-        4: ["%Y"],
-        6: ["%Y%m"],
-        7: ["%Y-%m"],
-        8: ["%Y%m%d"],
-        10: ["%Y%m%d%H", "%Y-%m-%d"],
-        12: ["%Y%m%d%H%M"],
-        19: ["%Y-%m-%dT%H:%M:%S"],
-    }
-
-    def _parse_date(date, fmts):
-        for fmt in fmts:
-            try:
-                s = datetime.strptime(date, fmt)
-            except ValueError:
-                pass
-            else:
-                match = fmt
-                break
-        else:
-            raise ValueError(f"Can't parse date {date} with formats {fmts}.")
-        return s, match
-
-    fmt = None
-    if isinstance(date, str):
-        try:
-            date, fmt = _parse_date(date, fmts[len(date)])
-        except (KeyError, ValueError):
-            date = pd.NaT
-    elif isinstance(date, cftime.datetime):
-        for n in range(3):
-            try:
-                date = datetime.fromisoformat((date - pd.Timedelta(n)).isoformat())
-            except ValueError:  # We are NOT catching OutOfBoundsDatetime.
-                pass
-            else:
-                break
-        else:
-            raise ValueError(
-                "Unable to parse cftime date {date}, even when moving back 2 days."
-            )
-    elif isinstance(date, pd.Timestamp):
-        date = date.to_pydatetime()
-
-    if not isinstance(date, pd.Period):
-        date = pd.Period(date, freq=freq)
-
-    if end_of_period and fmt:
-        if "m" not in fmt:
-            date = date.asfreq("A-DEC").asfreq(freq)
-        elif "d" not in fmt:
-            date = date.asfreq("M").asfreq(freq)
-        # TODO: Implement subdaily ?
-
-    if out_dtype == "str":
-        return date.strftime(strtime_format)
-
-    if out_dtype == "datetime":
-        return date.to_timestamp()
-    return date
 
 
 def _build_id(element: pd.Series, columns: list[str]):
